+++ libphx/src/CullFace.rs
@@ -17,15 +17,15 @@ pub unsafe extern "C" fn CullFace_Push(mut cullFace: CullFace) {
 
 #[no_mangle]
 pub unsafe extern "C" fn CullFace_PushNone() {
-    RenderState_PushCullFace(0 as i32);
+    RenderState_PushCullFace(0_i32);
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn CullFace_PushBack() {
-    RenderState_PushCullFace(1 as i32);
+    RenderState_PushCullFace(1_i32);
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn CullFace_PushFront() {
-    RenderState_PushCullFace(2 as i32);
+    RenderState_PushCullFace(2_i32);
 }
