+++ libphx/src/File.rs
@@ -19,9 +19,9 @@ pub unsafe extern "C" fn File_Exists(mut path: *const libc::c_char) -> bool {
     let mut f: *mut libc::FILE = libc::fopen(path, b"rb\0" as *const u8 as *const libc::c_char);
     if !f.is_null() {
         libc::fclose(f);
-        return 1 as i32 != 0;
+        return 1_i32 != 0;
     }
-    return 0 as i32 != 0;
+    return 0_i32 != 0;
 }
 
 #[no_mangle]
@@ -63,18 +63,18 @@ pub unsafe extern "C" fn File_ReadBytes(mut path: *const libc::c_char) -> *mut B
         return std::ptr::null_mut();
     }
     libc::fseek(file, 0 as libc::c_long, libc::SEEK_END);
-    let mut size: i64 = libc::ftell(file) as i64;
-    if size == 0 as i64 {
+    let mut size: i64 = libc::ftell(file);
+    if size == 0_i64 {
         return std::ptr::null_mut();
     }
-    if size < 0 as i64 {
+    if size < 0_i64 {
         Fatal(
             b"File_Read: failed to get size of file '%s'\0" as *const u8 as *const libc::c_char,
             path,
         );
     }
     libc::rewind(file);
-    if size > sdl2_sys::UINT32_MAX as u32 as i64 {
+    if size > sdl2_sys::UINT32_MAX as i64 {
         Fatal(
             b"File_Read: filesize of '%s' exceeds 32-bit capacity limit\0" as *const u8
                 as *const libc::c_char,
@@ -82,8 +82,8 @@ pub unsafe extern "C" fn File_ReadBytes(mut path: *const libc::c_char) -> *mut B
         );
     }
     let mut buffer: *mut Bytes = Bytes_Create(size as u32);
-    let mut result: usize = libc::fread(Bytes_GetData(buffer), size as usize, 1 as usize, file);
-    if result != 1 as usize {
+    let mut result: usize = libc::fread(Bytes_GetData(buffer), size as usize, 1_usize, file);
+    if result != 1_usize {
         Fatal(
             b"File_Read: failed to read correct number of bytes from '%s'\0" as *const u8
                 as *const libc::c_char,
@@ -100,12 +100,12 @@ pub unsafe extern "C" fn File_ReadCstr(mut path: *const libc::c_char) -> *const 
     if file.is_null() {
         return std::ptr::null();
     }
-    libc::fseek(file, 0 as libc::c_long, 2 as i32);
-    let mut size: i64 = libc::ftell(file) as i64;
-    if size == 0 as i64 {
+    libc::fseek(file, 0 as libc::c_long, 2_i32);
+    let mut size: i64 = libc::ftell(file);
+    if size == 0_i64 {
         return std::ptr::null();
     }
-    if size < 0 as i64 {
+    if size < 0_i64 {
         Fatal(
             b"File_ReadAscii: failed to get size of file '%s'\0" as *const u8
                 as *const libc::c_char,
@@ -115,11 +115,11 @@ pub unsafe extern "C" fn File_ReadCstr(mut path: *const libc::c_char) -> *const 
     libc::rewind(file);
     let mut buffer: *mut libc::c_char = MemAlloc(
         (::core::mem::size_of::<libc::c_char>())
-            .wrapping_mul((size as usize).wrapping_add(1 as usize)),
+            .wrapping_mul((size as usize).wrapping_add(1_usize)),
     ) as *mut libc::c_char;
     let mut result: usize =
-        libc::fread(buffer as *mut libc::c_void, size as usize, 1 as usize, file);
-    if result != 1 as usize {
+        libc::fread(buffer as *mut libc::c_void, size as usize, 1_usize, file);
+    if result != 1_usize {
         Fatal(
             b"File_Read: failed to read correct number of bytes from '%s'\0" as *const u8
                 as *const libc::c_char,
@@ -127,7 +127,7 @@ pub unsafe extern "C" fn File_ReadCstr(mut path: *const libc::c_char) -> *const 
         );
     }
     libc::fclose(file);
-    *buffer.offset(size as isize) = 0 as i32 as libc::c_char;
+    *buffer.offset(size as isize) = 0_i32 as libc::c_char;
     return buffer as *const libc::c_char;
 }
 
@@ -135,17 +135,17 @@ pub unsafe extern "C" fn File_ReadCstr(mut path: *const libc::c_char) -> *const 
 pub unsafe extern "C" fn File_Size(mut path: *const libc::c_char) -> i64 {
     let mut file: *mut libc::FILE = libc::fopen(path, b"rb\0" as *const u8 as *const libc::c_char);
     if file.is_null() {
-        return 0 as i32 as i64;
+        return 0_i32 as i64;
     }
-    libc::fseek(file, 0 as libc::c_long, 2 as i32);
-    let mut size: i64 = libc::ftell(file) as i64;
+    libc::fseek(file, 0 as libc::c_long, 2_i32);
+    let mut size: i64 = libc::ftell(file);
     libc::fclose(file);
     return size;
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn File_Read(mut this: *mut File, mut data: *mut libc::c_void, mut len: u32) {
-    libc::fread(data, len as usize, 1 as usize, (*this).handle);
+    libc::fread(data, len as usize, 1_usize, (*this).handle);
 }
 
 #[no_mangle]
@@ -154,7 +154,7 @@ pub unsafe extern "C" fn File_Write(
     mut data: *const libc::c_void,
     mut len: u32,
 ) {
-    libc::fwrite(data, len as usize, 1 as usize, (*this).handle);
+    libc::fwrite(data, len as usize, 1_usize, (*this).handle);
 }
 
 #[no_mangle]
@@ -162,7 +162,7 @@ pub unsafe extern "C" fn File_WriteStr(mut this: *mut File, mut data: *const lib
     libc::fwrite(
         data as *const libc::c_void,
         StrLen(data),
-        1 as usize,
+        1_usize,
         (*this).handle,
     );
 }
@@ -172,8 +172,8 @@ pub unsafe extern "C" fn File_ReadI64(mut this: *mut File) -> i64 {
     let mut value: i64 = 0;
     libc::fread(
         &mut value as *mut i64 as *mut libc::c_void,
-        ::core::mem::size_of::<i64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i64>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -183,8 +183,8 @@ pub unsafe extern "C" fn File_ReadI64(mut this: *mut File) -> i64 {
 pub unsafe extern "C" fn File_WriteU64(mut this: *mut File, mut value: u64) {
     libc::fwrite(
         &mut value as *mut u64 as *const libc::c_void,
-        ::core::mem::size_of::<u64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u64>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -193,8 +193,8 @@ pub unsafe extern "C" fn File_WriteU64(mut this: *mut File, mut value: u64) {
 pub unsafe extern "C" fn File_WriteI8(mut this: *mut File, mut value: i8) {
     libc::fwrite(
         &mut value as *mut i8 as *const libc::c_void,
-        ::core::mem::size_of::<i8>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i8>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -203,8 +203,8 @@ pub unsafe extern "C" fn File_WriteI8(mut this: *mut File, mut value: i8) {
 pub unsafe extern "C" fn File_WriteU32(mut this: *mut File, mut value: u32) {
     libc::fwrite(
         &mut value as *mut u32 as *const libc::c_void,
-        ::core::mem::size_of::<u32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u32>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -213,8 +213,8 @@ pub unsafe extern "C" fn File_WriteU32(mut this: *mut File, mut value: u32) {
 pub unsafe extern "C" fn File_WriteI16(mut this: *mut File, mut value: i16) {
     libc::fwrite(
         &mut value as *mut i16 as *const libc::c_void,
-        ::core::mem::size_of::<i16>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i16>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -224,8 +224,8 @@ pub unsafe extern "C" fn File_ReadU64(mut this: *mut File) -> u64 {
     let mut value: u64 = 0;
     libc::fread(
         &mut value as *mut u64 as *mut libc::c_void,
-        ::core::mem::size_of::<u64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u64>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -236,8 +236,8 @@ pub unsafe extern "C" fn File_ReadI32(mut this: *mut File) -> i32 {
     let mut value: i32 = 0;
     libc::fread(
         &mut value as *mut i32 as *mut libc::c_void,
-        ::core::mem::size_of::<i32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i32>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -248,8 +248,8 @@ pub unsafe extern "C" fn File_ReadU32(mut this: *mut File) -> u32 {
     let mut value: u32 = 0;
     libc::fread(
         &mut value as *mut u32 as *mut libc::c_void,
-        ::core::mem::size_of::<u32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u32>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -259,8 +259,8 @@ pub unsafe extern "C" fn File_ReadU32(mut this: *mut File) -> u32 {
 pub unsafe extern "C" fn File_WriteU16(mut this: *mut File, mut value: u16) {
     libc::fwrite(
         &mut value as *mut u16 as *const libc::c_void,
-        ::core::mem::size_of::<u16>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u16>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -270,8 +270,8 @@ pub unsafe extern "C" fn File_ReadU8(mut this: *mut File) -> u8 {
     let mut value: u8 = 0;
     libc::fread(
         &mut value as *mut u8 as *mut libc::c_void,
-        ::core::mem::size_of::<u8>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u8>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -282,8 +282,8 @@ pub unsafe extern "C" fn File_ReadU16(mut this: *mut File) -> u16 {
     let mut value: u16 = 0;
     libc::fread(
         &mut value as *mut u16 as *mut libc::c_void,
-        ::core::mem::size_of::<u16>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u16>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -294,8 +294,8 @@ pub unsafe extern "C" fn File_ReadI16(mut this: *mut File) -> i16 {
     let mut value: i16 = 0;
     libc::fread(
         &mut value as *mut i16 as *mut libc::c_void,
-        ::core::mem::size_of::<i16>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i16>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -306,8 +306,8 @@ pub unsafe extern "C" fn File_ReadF32(mut this: *mut File) -> f32 {
     let mut value: f32 = 0.;
     libc::fread(
         &mut value as *mut f32 as *mut libc::c_void,
-        ::core::mem::size_of::<f32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<f32>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -318,8 +318,8 @@ pub unsafe extern "C" fn File_ReadF64(mut this: *mut File) -> f64 {
     let mut value: f64 = 0.;
     libc::fread(
         &mut value as *mut f64 as *mut libc::c_void,
-        ::core::mem::size_of::<f64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<f64>(),
+        1_usize,
         (*this).handle,
     );
     return value;
@@ -329,8 +329,8 @@ pub unsafe extern "C" fn File_ReadF64(mut this: *mut File) -> f64 {
 pub unsafe extern "C" fn File_WriteU8(mut this: *mut File, mut value: u8) {
     libc::fwrite(
         &mut value as *mut u8 as *const libc::c_void,
-        ::core::mem::size_of::<u8>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<u8>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -339,8 +339,8 @@ pub unsafe extern "C" fn File_WriteU8(mut this: *mut File, mut value: u8) {
 pub unsafe extern "C" fn File_WriteI32(mut this: *mut File, mut value: i32) {
     libc::fwrite(
         &mut value as *mut i32 as *const libc::c_void,
-        ::core::mem::size_of::<i32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i32>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -349,8 +349,8 @@ pub unsafe extern "C" fn File_WriteI32(mut this: *mut File, mut value: i32) {
 pub unsafe extern "C" fn File_WriteI64(mut this: *mut File, mut value: i64) {
     libc::fwrite(
         &mut value as *mut i64 as *const libc::c_void,
-        ::core::mem::size_of::<i64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i64>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -359,8 +359,8 @@ pub unsafe extern "C" fn File_WriteI64(mut this: *mut File, mut value: i64) {
 pub unsafe extern "C" fn File_WriteF32(mut this: *mut File, mut value: f32) {
     libc::fwrite(
         &mut value as *mut f32 as *const libc::c_void,
-        ::core::mem::size_of::<f32>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<f32>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -369,8 +369,8 @@ pub unsafe extern "C" fn File_WriteF32(mut this: *mut File, mut value: f32) {
 pub unsafe extern "C" fn File_WriteF64(mut this: *mut File, mut value: f64) {
     libc::fwrite(
         &mut value as *mut f64 as *const libc::c_void,
-        ::core::mem::size_of::<f64>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<f64>(),
+        1_usize,
         (*this).handle,
     );
 }
@@ -380,8 +380,8 @@ pub unsafe extern "C" fn File_ReadI8(mut this: *mut File) -> i8 {
     let mut value: i8 = 0;
     libc::fread(
         &mut value as *mut i8 as *mut libc::c_void,
-        ::core::mem::size_of::<i8>() as usize,
-        1 as usize,
+        ::core::mem::size_of::<i8>(),
+        1_usize,
         (*this).handle,
     );
     return value;
