+++ libphx/src/GUID.rs
@@ -3,7 +3,7 @@ use glam::Vec3;
 use libc;
 
 #[no_mangle]
-pub static mut nextID: u64 = 1 as i32 as u64;
+pub static mut nextID: u64 = 1_i32 as u64;
 
 #[no_mangle]
 pub unsafe extern "C" fn GUID_Create() -> u64 {
@@ -14,10 +14,10 @@ pub unsafe extern "C" fn GUID_Create() -> u64 {
 
 #[no_mangle]
 pub unsafe extern "C" fn GUID_Exists(mut id: u64) -> bool {
-    return id < nextID && id != 0 as u64;
+    return id < nextID && id != 0_u64;
 }
 
 #[no_mangle]
 pub unsafe extern "C" fn GUID_Reset() {
-    nextID = 1 as i32 as u64;
+    nextID = 1_i32 as u64;
 }
